Plugins: Transform requests that modify the DB to POST (#418)
* CSRF plugin: make delete a POST (blacklist_number)
* CSRF plugin: tranform DB updates to POST (server_alert)
* CSRF plugin: tranform DB updates to POST (sms_credit)
* CSRF plugin: tranform DB updates to POST (sms_to_wordpress)
* CSRF plugin: tranform DB updates to POST (sms_to_xmpp)
* CSRF plugin: tranform DB updates to POST (soap)
* CSRF plugin: tranform DB updates to POST (stop_manager)
* CSRF plugin: tranform DB updates to POST (whitelist_number)
* plugin: fix redirect target

+++ application/plugins/blacklist_number/controllers/Blacklist_number.php
@@ -59,9 +59,12 @@ class Blacklist_number extends Plugin_controller {
 		$this->load->view('main/layout', $data);
 	}
 
-	function delete($id)
+	function delete()
 	{
-		$this->blacklist_number_model->delete($id);
-		redirect('plugin/blacklist_number');
+		if ($_POST)
+		{
+			$id = intval($this->input->post('id'));
+			$this->blacklist_number_model->delete($id);
+		}
 	}
 }
