+++ application/plugins/whitelist_number/controllers/Whitelist_number.php
@@ -58,9 +58,12 @@ class Whitelist_number extends Plugin_controller {
 		$this->load->view('main/layout', $data);
 	}
 
-	function delete($id)
+	function delete()
 	{
-		$this->whitelist_number_model->delete($id);
-		redirect('plugin/whitelist_number');
+		if ($_POST)
+		{
+			$id = intval($this->input->post('id'));
+			$this->whitelist_number_model->delete($id);
+		}
 	}
 }
